Explicitly name base class for dependent names
Circa 2009 when nocycle was written, when you used a template parameter
as a base class it might let you use inherited methods from that base
class without any kind of qualification at the callsite.  That is now
explicitly prohibited by the standard:

https://isocpp.org/wiki/faq/templates#nondependent-name-lookup-members

This adds qualifiers to solve errors related to that, and also changes
the base class name to the shorter "Base" instead of "BaseClass".