Fix CMake condition, fit in 80 columns
There was a mismatching end condition label which old versions of
CMake apprently considered commentary and did not check...but now
causes warnings if you are in "developer mode".

This resolves that and generally formats the CMake file to be more
friendly for viewing online (80 column margin, better spacing).  Also
updates links and removes mention of `cmake -i`, as the cmake tool no
longer has an interactive mode (one has to install a separate "ccmake")