Remove unused boost graph code, which now causes errors
Boost 1.55 introduced a change which caused some unused code (that was
not useful) to have a compile time error informing that it was not
useful:

https://www.boost.org/users/history/version_1_55_0.html

    "void is no longer allowed as a bundled property type (for example,
    in the VertexProperties template parameters to graph types); it
    did not work reliably before, but a static assertion now forbids it
    entirely. Use boost::no_property instead."

The generic algorithms (such as depth first search) want to be able to
keep track of data associated with things like edges or vertices, but
don't want to mandate how or where this data is stored.  If you have
room in your graph structure itself, you can put it there; or you can
point it to some associative maps you've already defined.  Then if
you don't provide anything, it will create its own sidestructure.

There was a declaration for a `colorMap` which was a property map
that implied some kind of storage was in the adjacency_matrix itself,
when no such storage had been defined.  Instead, all the calls would
either accept the default or pass one in on the fly.  So there wasn't
any compile-time checking of the poorly-formed unused definition that
falsely suggested the graph had been defined with its own storage.

This happened because the Boost.Graph reference implementation was a
2009-era copy and paste from sporadic usenet sources.  (This is not to
say that good examples of use of the Boost.Graph library are all that
easy to find in 2018--it's a bit esoteric and not many people know how
to use it properly.)  But also because the sources were compiled
without high warning levels that would notice an unused declaration.