Temporarily use Nstate library in nocycle, not submodule
It seems that some work was done on the Nstate library independent of
nocycle circa 2012, to see if it could become a header-only library,
due to advances in C++11.

While my memory is hazy on what motivated that 6 years ago, the change
looks to have involved only messing around with nstate--while not
making sure nocycle could deal with it.  The compromise appears to have
been to snapshot Nstate into local files in the nocycle directory, with
a likely intent to come back to it later--if it ever was needed.

As a temporary measure this uses the copy of Nstate that was cached in
nocycle, just to get it to build.  Resolving it will require updating
the build system and figuring out if Nstate can (should?) be packaged
as an independent project...that people interested in packed arrays
can use without also dealing with the graph stuff.

Also adds a missing cassert include needed for compilation.