Remove NocycleSettings.hpp
NocycleSettings.hpp is a holdover from a pre-CMake era in which the
configuring of nocycle was done by editing a header file.  Then it
was changed to generate the header file (from a ".in" version as
a template), so the file is not currently used.

So this deletes it.  But more broadly, the header generation was really
just an early experiment with CMake to try and learn it in 2009...and
to make a small incremental improvement to the "configurability".  It
should probably be done entirely differently, if there's reason enough
to be improving it.

Also adds quotes to the parameter for the `#error` directives in the
NocycleConfig.hpp.in, because otherwise you can't use apostrophes :-(

https://stackoverflow.com/a/37195204/211160