Remove redundant `#pragma once` and include guards
It seems fairly pointless to have *both* `#pragma once` and include
guards in the modern era, but even fairly well known libraries like
boost would do it.  It is based on the belief that compilers that do
support `#pragma once` will have some kind of more efficient logic to
being able to avoid scanning the entire file, while the fallback will
be available to older compilers.

https://stackoverflow.com/q/50090382

For nocycle's purposes, there is no particular reason to not embrace
`#pragma once`, as it is widely available and "nearly standard" in
modern C++ compilers...(while not being *actually* in the standard).