Eliminate nocycle_assert() in favor of plain assert()
While many build environments combine `assert()` with having debug
symbols, and not having it with not having them, it's possible to
control this in C with defining NDEBUG.  In 2009 I may or may not have
known that...but it's also possible that I was considering using
a custom assert for other reasons.

This simplifies and just uses regular assert() in the code; if finer
grained control than NDEBUG is needed, that should be done with another
conditional compilation flag.