Fix out of bounds access to power table
The SetDigitInPackedValue() routine was asking for the power for the
digit plus one, even in cases where it did not use it.  This led to
an out of bounds access in the power table.  The value was not used,
so it did not cause a noisy failure at runtime, but could still be
a problem.

Added an assert to PowerForDigit() to check this condition.